boot: add support for python properties oracle
This patch adds support for building a snapshot of the dCache 
configuration as a Python oracle class.  This is similar to the 
existing support for building a POSIX shell function for querying 
property values.  The output is triggered by calling the bootloader's 
'compile' command with the '-python' option.  The quiet option '-q' is 
recommended.

The motivation is to allow Python scripts to react to dCache 
configuration properties.  The output is designed to be executed by 
the Python script using Python's builtin command 'exec'.  This results 
in defining the class Properties, which is instantiated as the 
variable properties.  The class has a single method 'get' that allows 
querying of property values as global, domain and service scopes.  The 
number of arguments determines the scope: a single argument is global 
scope, two arguments is domain scope and three is for service scope.  
If a property is not defined then None is returned.

Here is a sample Python script that demonstrates this in action.  This 
example uses the dcache script to invoke the boot-loader, which it 
assumes is on the path.

    #!/usr/bin/env python

    from subprocess import Popen, PIPE
    d = Popen(["dcache", "loader", "-q", "compile", "-python"], stdout=PIPE)
    exec d.communicate()[0]

    #  Print the value of a globally defined property
    print properties.get('info-provider.se-unique-id')

    #  Print the value of a property for a specific domain
    print properties.get('dcache.java.memory.heap', 'dCacheDomain')

    #  Print the value of a property for a specific service-instance
    print properties.get('path', 'dCacheDomain', 'pool1')
    print properties.get('path', 'dCacheDomain', 'pool2')


Target: trunk
Require-book: no
Require-notes: no
Patch: http://rb.dcache.org/r/4938/
Acked-by: Gerd Behrmann