pnfsmanager: add option to verify lookup permissions of entire path
Moves the permission check previously performed in
PermissionHandlerNameSpaceProvider into ChimeraNameSpaceProvider. This
allows us to reduce the overhead of those checks by avoiding repeated
reads of the same attributes.

PermissionHandlerNameSpaceProvider is still used for PNFS:
PnfsNameSpaceProvider is introduced as a helper class to allow the two
providers to have the same interface.

There is room for further improvements in some of the checks.

This patch provides the foundation for another patch that adds an
optional lookup permission check for the complete path and in the future
for dynamic ACL inheritance.

This patch adds the path2inodes method. It too resolves the path, but it
returns the inodes of all path elements, not just the final inode. This
is quite useful when checking lookup permissions, as we need the inodes
of the path elements too.

The stat cache of the inode is prepulated. Since we have to access the
record in t_inodes anyway, returning the data to the client adds minimal
overhead and avoids that we have to fetch the stat data for the inodes
afterwards.

For postgresql a stored procedure resolves the complete path and returns
a set of inodes in a single roundtrip to the database.

Support added for Chimera for complete path lookup.

When ACL support is disabled, the check should not add any
noticable overhead.  When both full-path-checking and ACLs are
enabled, the check will trigger several additional round trips to
the database to fetch the ACLs.

RELEASE NOTES:
Chimera users must recreate the stored procedures by applying
pgsql-procedures.sql.

Require-book: yes
Require-notes: yes
Merged: trunk@15551
Merged: trunk@15567
Merged: trunk@15568
Patch: http://rb.dcache.org/r/3214
Patch: http://rb.dcache.org/r/3216
Patch: http://rb.dcache.org/r/3217