missingfiles: add support
Currently, when a user requests to read a file that doesn't exist, 
the door will respond to the user with the appropriate error message.

There are two use-cases where something more sophisticated is 
desirable.  The first is that dCache should notify one or more 
external services that it was requested to read a file that doesn't 
exist.  The second is that dCache will populate the missing file from 
some external source, allowing the read request to succeed.  There 
may be other examples where it is desirable for dCache to react to 
failed read attempts.

The MissingFiles service is a new dCache component.  It exists to 
allow dCache to react in a coordinated fashion to a user's request to 
read a file that doesn't exist.  This central service instructs the 
door to either fail the request or retry (which makes sense only if 
the file has been fetched from some external source).  Doors need to 
be updated to support this; however, the amount of additional code in 
the door should be small.  To demonstrate this, the WebDAV door is 
updated to provide this support.

The MissingFiles service is pluggable.  It takes a list of plugins 
and instantiates them.  These plugins are used to determine how 
dCache should react when a user attempts to read a missing file.  
Each plugin is asked in turn what to do until a plugin replies with a 
terminating answer or the list of plugins is exhausted.  A plugin 
replies saying to fail the request, to retry the request or to ask 
the next plugin in the chain.  If the last plugin defers the request 
then then the MissingFiles service will instruct the door to fail the 
request.

To demonstrate the plugins, this patch includes a simple notification 
plugin that sends a message for each notified file.

With the default configuration, the WebDAV door does not send 
messages to the MissingFiles service; therefore, sites should be 
unaffected unless they enable support.

Note that populating dCache with files from a remote storage system 
is currently not possible since the plugin interface does not allow 
triggering 3rd-party copies.  A future patch will fix this.

Target: trunk
Patch: http://rb.dcache.org/r/4879/
Acked-by: Tigran Mkrtchyan