Logging-based alarms, part 4 (tests)
This is part 4 (out of 4) of the third version of the proposed alarms component.  

This patch provides unit tests for the logging, DAO and Wicket classes found in:

    http://rb.dcache.org/r/4885
    http://rb.dcache.org/r/4888
    http://rb.dcache.org/r/4890

The unit tests are written in BDD style and use mock-ito for the storage interfaces.

In addition, there is an alarm simulator class which can be used together with the logback-test.xml provided in modules/dcache/src/tests/resources for sending mock checksum alarms to a remote service (most easily done inside an IDE such as Eclipse).  The command-line options set the number of separate threads to use (-threads), the max timeout interval between iterations (-maxTimeoutInSeconds), and the frequency with which to produce an error (per teration per thread: -errorFrequency). The timeout is randomized within 1 second and max for any given thread. Each simulator thread reports on one file checksum; this is so handling of duplicate errors can be tested on the back-end.

All Junit tests were run.   The four patches were applied to trunk and deployed to a testbed installation.  The logback-server was booted and the ChecksumAlarmSimulator was run as client; the server logs, alarms.xml and Alarm pages were observed to be correct, particularly in handling duplicate events.  

Several files in a pool where modified and the admin command csm set policy -scrub=on was issued; the checksum mismatch error was reported and the alarm arrived at the server and was displayed in the web page.


Target: trunk
Patch: http://rb.dcache.org/r/4892
Acked-by: Dmitry
Require-notes: no
Require-book: no