Allow for dynamic creation of XML storage file for alarms
The current version of alarms relies on the existence of the alarms.xml provided through the rpm build in the known location (/var/lib/dcache/alarms) .   The default setup, however, causes system-tests to fail because the webadmin alarm cleanup thread is also activated and cannot find that file in the tree used for that setup.  

While this situation can be easily remedied by turning off (disabling) the cleanup thread, the more general case remains user-unfriendly.

This patch adds protective code which will attempt to create the alarms.xml file if it does not exist and if the path is defined and points to a writable directory.  

The cleaner thread is also disabled in dcache.conf for system-tests.

The webadmin.alarm.xml.path property has also been changed to the more functionally descriptive webadmin.alarm.store.path.

Testing: Redeployed, testing with no alarms.xml in the defined path, and also with the path set to a non-default location.

Target: trunk
Patch: http://rb.dcache.org/r/4973
Acked-by: Dmitry
Requires-notes: no
Requires-book: yes (only insofar as the property name has changed)