pool: Add flag to migration {copy/move/cache} to override the disable state of the source pool
When a pool is disabled due to some sort of failure, there may exist a need to evacuate the files
located on the pool, especially if they are the only replica of that particular file.

Re-enabling the pool has the problem of allowing external requests and has the risk of re-disabling
the pool in the process. The -force-source-mode flag to these commands is routed through to the
source pool in the p2p messages, allowing for an override of the guard for transmitting from a
disabled pool.

Target: trunk
Require-notes: yes
Require-book: yes
Ticket: http://rt.dcache.org/Ticket/Display.html?id=7464
Patch: http://rb.dcache.org/r/4981
Acked-by: Tigran
Acked-by: Karsten