add RemoteNameSpaceProvider
This patch introduces the RemoteNameSpaceProvider class.  
RemoteNameSpaceProvider implements the NameSpaceProvider interface by 
sending messages to PnfsManager.  The bulk of this work is already 
contained in the PnfsHandler class, so RemoteNameSpaceProvider is 
relatively light-weight.

There was some functionality that NameSpaceProvider makes available 
that is not currently available via dCache messages.  This patch adds 
an additional message (PnfsRemoveChecksumMessage) and updates an 
existing message (PnfsMapPathMessage) to provide this missing 
functionality.

By introducing this class, components that currently use PnfsHandler 
(e.g., pool, doors, ..) may be able to use the NameSpaceProvider 
interface and operating against an instant of RemoteNameSpaceProvider 
instead.  Such components then become agnostic of how the namespace 
operations are achieved.  This, in turn, would allow such components to 
adopt different strategies; for example, talking directly to Chimera or 
become part of a distributed namespace.

It is possible that some PnfsHandler operations are available that do 
not correspond to NameSpaceProvider methods.  In such cases, 
NameSpaceProvider interface may need to be adjusted to allow doors to 
use RemoteNameSpaceProvider.  This corresponds (roughly) to moving 
functionality from PnfsManager into the namespace implementations 
(currently just ChimeraNameSpaceProvider).

Some effort was spent in providing unit-tests for this class.  This is 
to allow future refactoring of this class.  For example, once 
RemoteNameSpaceProvider is the only class using PnfsHandler then the 
two classes may be refactored.

Target: trunk
Require-notes: no
Require-book: no
Patch: http://rb.dcache.org/r/4932/
Acked-by: Tigran Mkrtchyan