gplazma: allow plugins to be 'dCache aware'
gPlazma is designed to be a stand-alone component that dCache happens 
to use for authentication and identitiy management.  The design is 
such that other software projects may use gPlazma; therefore, plugins 
currently are unaware that they are being used by a dCache system.

This patch allows for plugins that interact with dCache.  The 
immediate use-case for this is to have a plugin that verifies a user's 
home directory exists.  Such a plugin could even attempt to create the 
home directory if it finds it doesn't already exist in the namespace.

The patch has two parts: updating gPlazma to allow plugin decoration 
and a plugin decorator that allows plugins to be 'namespace aware'.  
The first part (allows plugin decoration) is achieved by the gPlazma 
class accepting a custom PluginFactory implementation.  The second 
part is update dCache's use of gPlazma to include a custom 
PluginFactory that looks for plugins that declare themselves 
NamespaceAware, which are given a NameSpaceProvider (in fact, a 
RemoteNameSpaceProvider) with which to manipulate dCache's namespace.

Acked-by: Tigran Mkrtchyan
Patch: http://rb.dcache.org/r/4933/
Target: trunk
Require-notes: no
Require-book: no