gsidcap: remove extra TCP frame
When investigating a problem with gsidcap, I noticed that the encoded 
messages were sent as two TCP frames: the first being content bearing 
("enc " + base64-encoded data) the second containing the single 
carriage-return character ("\n").  This happens because LoginManager 
switches off Nagle's algorithm (so we are more responsible for TCP 
framing) and the GSI tunnel writes the output in two parts.

This is inefficient as it introduces an additional communication 
overhead.  However, the effect is likely to be slight since most 
gsidcap communication is spent sending or receiving data.

This patch removes this inefficiency by composing the complete message 
and sending in one go.

FOR RELEASE NOTES:

Include a simple statement that the network output for gsidcap has 
been slightly optimised by sending more data in one go on the control 
channel.  That this will likely have little effect as, for majority of 
dcap transfers, most of the time is spend sending or receiving data.

Target: trunk
Require-nodes: yes
Require-book: no
Patch: http://rb.dcache.org/r/5040
Acked-by: Tigran Mkrtchyan