gsi-/keberos-dcap: don't sent '\n' as separate SSL frame
The gsidcap and the kerberosdcap door replies on the control channel 
using encrypted data.  The DCAP protocol ends each respond with a 0x0a 
0x0d two-byte sequence.  Currently the text of the response plus the 
first byte (0x0a) is send as a single GSS-encoded frame and the final 
0x0d byte in its own, separate frame.

Apart from the overhead of an unnecessary SSL frame, when a block 
cipher is used an additional padding is needed.  When testing, I found 
that 23 characters were added as padding.  The result is that sending 
the 0x0d byte separately adds some 29 bytes of text (this includes the 
SSL frame).  When base64-encoded the and placed in the DCAP frame, 
this is about 40 bytes of text.

The problem arises from TunnelOutputStream use of an implicit flush at 
the end-of-line.  This currently happens when there is a 0x0a or a 
0x0d character.  The 0x0a 0x0d two-byte sequence is, thus, prematurely 
flushed.

This patch fixes this problem by switching TunnelOutputStream to use 
explicit flushing.  The DCAP door already sends an explicit flush 
after each response line.  However, due to a bug in TelnetStreamEngine 
/ TelnetOutputStream2, this flush is not propagated to the Tunnel. 
Therefore, the patch also fixes this problem.

RELEASE NOTES:

Mention that the gsidcap and kerberosdcap responses on the control 
channel are now sent in a more optimal fashion.  This is unlikely to 
improve the performance of gsidcap or kerberosdcap much since, for 
most transfers, most of the time is spend sending or receiving data on 
the data channel.

Target: trunk
Require-notes: yes
Require-book: no
Patch: http://rb.dcache.org/r/5041
Acked-by: Tigran Mkrtchyan