gplazma: add missing directory-listing-handler
During the review process, Gerd pointed out that some of functionality 
of the RemoteNameSpaceProvider was already provided with the 
DirectoryListingHandler.  The patch was refactored to use this, but a 
subsequent patch that makes use of RemoteNameSpaceProvider was 
mistakenly not updated accordingly.  The effect is that the dCache 
domain that hosts gPlazma does not start.

This patch fixes this by creating a DirectoryListHandler, which is 
injected into the RemoteNameSpaceProvider as the required constructor 
argument.

Target: trunk
Request: 2.5
Patch: http://rb.dcache.org/r/5039
Acked-by: Tigran Mkrtchyan