chimera-cli: fix output of 'ls' cmd
The chimera-cli may be used as a diagnostic command for investigating 
problems with dCache's namespace.  In particular, it may be used by an 
admin who is new to dCache to "see what's happening".

Although the current output presents all the necessary information the 
format is alien.  This is quite off-putting for the new admin.

This patch changes the output to follow what 'ls -l' shows.  It also 
forces '.' and '..' to appear as the first and second items 
(respectively).  Although this is purely cosmetic, I think it helps 
the output look more familiar.

FOR RELEASE NOTES

Output from chimera-cli ls command has been updated to match 'ls -l'

Target: trunk
Request: 2.5
Require-book: yes
Require-notes: yes
Patch: http://rb.dcache.org/r/5037