scripts: add state duration to 'dcache status' output
The 'dcache status' command shows in which state each are of the 
domains: stopped, running, orphaned or restarting.

Unfortunately, if a domain is recently restarted then this isn't 
obvious since the state is, once again, 'running'.  Therefore, one has 
to be lucky (by running the 'dcache status' command when the domain is 
sleeping between restarts) to notice the problem.

This patch adds information on how long a domain has been running or 
restarting.  The information is given with course-grain since it's only 
a simple indicator.  A single integer with a time-unit is shown.  The 
duration is given in seconds (if less than 2 minutes), minutes (if less 
than 2 hours), hours (if less than 2 days), days (if less than 2 weeks) 
or weeks otherwise.

This should allow an admin to spot easily any domain that is 
restarting.

FOR BOOK:

Suggest searching through all references of the 'dcache status' command 
and updating the output.

It may be useful to describe the output somewhere within the book.


FOR RELEASE NOTES:

Mention that when a domain is started or restarting that the 'dcache 
status' command now also prints for how long the domain has been in 
that state.

Target: trunk
Patch: http://rb.dcache.org/r/5052
Require-book: yes
Require-notes: yes