Backoff utilities: controller, algorithm interface and standard implementation
This utility provides a wrapper class for running a task with back-off retry.

The controller encapsulates the CALL->WAIT/QUIT logic; it takes a Callable which returns a simple SUCCESS/FAILURE value, and an algorithm which computes the next wait time and whether the controller loop should quit.  

The Standard implementation is that of exponential back-off (based on radix 2).  Algorithms are provided via an implementation-specific factory interface.

The utility is internal and requires no user documentation.  Its use in new features should figure as part of the settable properties for that feature.

Target: trunk
Patch: http://rb.dcache.org/r/4977
Acked-by: Paul, Dmitry
Requires-notes: no
Requires-book: no