scripts: fix dcache cmd for zero-length cache file
When the /var partition becomes full, it can happen (because it happened 
to me) that running the 'dcache' command creates a zero-length cache 
file: /var/lib/dcache/config/cache.  Subsequent attempts to run the 
dcache command will fail with the following error:

[discordia] ~ # dcache status
/usr/share/dcache/lib/loadConfig.sh: line 89: getProperty: command not found
/usr/share/dcache/lib/loadConfig.sh: line 90: getProperty: command not found
/usr/bin/dcache: line 370: getProperty: command not found

The problem here is that the loadConfig.sh script does not recognise an 
empty file as a reason to regenerate the oracle, so doesn't (attempt to) 
repopulate the cache.

This patch fixes the problem by using the 'is not empty' test (-s) 
instead of the current 'exists' (-e) test.  This also makes the test 
more specific, so only files pass (thus excluding directories).

With this patch, to use the cached oracle, /var/lib/dcache/config/cache 
must exist, be a file and be non-empty.

Target: trunk
Request: 2.5
Request: 2.4
Patch: http://rb.dcache.org/r/5036/
Acked-by: Tigran Mkrtchyan
Acked-by: Dmitry Litvintsev