ssh2: Introduction of admin group ID
This patch allows to take into account the group ID of a user defined in the dcache.kpwd file
so that not all users can access the dCache ssh2 admin interface who are in the kpwd file.

There is a new property that makes the group configurable. It is admin.group in admin.properties
and its default value is 0.

Patch: http://rb.dcache.org/r/4714/
Acked-by: Karsten
Target: trunk
Request: 1.9.12
Request: 2.2

Merge-req:7569

RELEASE NOTES:
Admin users' group ID has to be 0 in order to administrate dCache when using the gPlazma2 KPWD 
plugin for authentication.
DOC:
The users who shall be enabled to administrate dCache have to have the GID 0 associated in the kpwd file.