srm: reduce dependency on SrmDCacheConnector
The SrmDCacheConnector class allows the Axis classes to discover the 
SRM instance, the Configuration instance, the back-end Storage 
instance.  This works by singleton anti-patterns using static class 
members and reflection.  It also requires that the Storage backend 
starts before the first Axis call, as the backend initialises the SRM, 
so populating the singleton instance.

Such approach is undesirable due to the fragile nature of the 
initialisation process, the use of reflection and the difficulties 
faced when testing classes that implement the singleton pattern.

Since future work will access the SRM instance from the spring 
framework.  The desire is to inject the three necessary objects 
(Configuration, SRM and Storage) directly, rather than via the 
SrmDCacheConnector class.

As a first step to removing SrmDCacheConnector completely, this patch 
refactors the Axis server objects to reduce their usage of 
SrmDCacheConnector.  In doing so, the SrmAuthorizer class was tidied up 
slightly to make it more readable.

Patch: http://rb.dcache.org/r/5076
Acked-by: Dmitry Litvintsev
Target: trunk
Require-notes: no
Require-book:no