srm: remove dead configuration options, fix 'gplazma' support
The SRM server has a Configuration class that holds various options 
that affect how the SRM behaves.

By using netbeans find-usages facility, I was able to identify those 
options who's getter method was never called.  If the getter isn't 
called then that option must have no effect on dCache code (modulo 
access via reflection).  This patch removes those options that have no 
effect on the SRM.  Almost all of them are not exposed to the admin, so 
this patch mostly has no effect.  The exception is the 
srmLsRequestLifetime dCache configuration property.  This patch marks 
this property as obsolete.

The Storage class (dCache-specific interface to the SRM) took a 
CellStub to a pool; however, this is never used.  This patch removes 
it.

The Storage class has a reference to the gPlazma cell as the hard-coded 
value "gPlazma".  There is a configuration property ("gplazma") which 
controls the name of the gPlazma cell a service should contact.  This 
patch removes the hard-coded name and replaces it with code that 
honours the gplazma property.

Finally, the SRM instance is provided as a singleton, with a static 
field member to hold a reference to this object.  This is, in part, to 
allow the code running from Axis to discover these objects (through 
reflection!).  This patch removes some dead-wood code in the method the 
Axis-code uses to discover the SRM instance -- code which was left over 
from when SRM ran within tomcat.

Some unnecessary Boolean object creation was fixed and some @Override 
annotations were added.

RELEASE NOTES:

Mention that the srmLsRequestLifetime property, which was never 
supported, has now been dropped.  The SRM now honours the 'gplazma' 
property.

Target: trunk
Require-notes: yes
Require-book: no
Patch: http://rb.dcache.org/r/5075
Acked-by: Dmitry Litvintsev