chimera-cli: add ability to remove a directory tag
The chimera-cli allows creating and querying directory tags but not 
deleting them.  In part, this is an inherited limitation from PNFS; 
however, the underlying database code has partial support for deleting 
tags (for deleting all tags in a directory).

This patch adds support in the database layer for deleting a single tag 
within a directory.  A new chimera-cli command is added that uses this 
new method to remove a tag.

FOR RELEASE NOTES:

Simple statement that tags may be deleted using chimera-cli

FOR BOOK:

Review where tags are described and update documentation to describe 
how to delete a tag.

Target: trunk
Require-notes: yes
Require-book: yes
Patch: http://rb.dcache.org/r/5050
Acked-by: Tigran Mkrtchyan