LocationManager: fix stack-trace when broker.host is unknown
When the admin configures broker.host with a host that isn't known 
dCache logs an error message and refuses to start.  Exceptions thrown 
during running a cell's constructor are propagated back to the 
CellShell that instantiates all the services in the domain, which 
prints a summary of the problem along with the filename and line number 
where the problem was triggered.

Unfortunately LocationManager currently also prints a stack trace 
whenever any Exception is thrown.  This means that mis-configuring 
LocationManager by not using an integer value or by configuring unknown 
hostname results in a stack-trace.

This is undesirable since stack-traces indicate a bug in dCache and 
this behaviour is not a bug.

This patch updates the behaviour so that exceptions that indicate 
admin-error are propagated without printing a stack-trace and a 
stack-trace is printed for those that are due to a bug in dCache (i.e., 
other subclasses of RuntimeException).

Target: trunk
Requires-book: no
Requires-notes: no
Patch: http://rb.dcache.org/r/5060/
Acked-by: Christian Bernardt