config: fix error message for obsolete properties
dCache configuration properties may be annotated to carry additional 
information; for 'obsolete' and 'forbidden' annotations, the property 
value becomes the error message returned to the user (if empty then a 
default message is used).

There are a few properties that were marked as obsolete have been left 
with their default values.  This means that the helpful error- or 
warning-message becomes baffling, since the message shows only the name 
of the property and the default value.

This patch fixes this by providing useful messages based on the 
comments immediately above the declaration.

The patch also moves the obsolete property declarations to the 
httpd.properties file so that the webadmin.properties file may be 
deleted.

The properties are updated to be forbidden.  Given that the properties 
used to affect port numbers and aspects of security, I felt it was 
unsafe to allow the domain to start if dCache still sees such 
configuration, which now has no effect.

Target: trunk
Request: 2.5
Patch: http://rb.dcache.org/r/5051/
Require-notes: no
Require-book: no