chimera-cli: fix chown so gid-ownership may be set
Chimera-cli has a chown command to alter the ownership of a file or 
directory.  Currently only the user-ownership may be altered with this 
command.  Chimera-cli also offers the chgrp command for altering the 
group-ownership of a file or directory.

This command updates the chown command so it supports setting gid 
ownership too.  This makes it more in-keeping with the GNU coreutils, 
which provides similar capability.

Target: trunk
Require-notes: yes
Require-book: yes
Patch: http://rb.dcache.org/r/5047/
Acked-by: Tigran Mkrtchyan