srm: inject SRM Configuration and Storage objects into Axis
The SRM runs within the Axis framework, which is instantiated by Spring 
(well, Jetty, really).  There are three key objects: SRM (the engine) 
Storage (the dCache-specific plugin) and Configuration (data object 
holding configuration).

When SRM code is running within the Axis framework, it must obtain 
access to these three objects to allow correct functionality.  The 
current mechanism that allows rather Heath Robinson.

A JNDI environment is used to associate the key "srmConfigFile" with a 
String that is the path to the srmConfigFile.xml file.

When a SRM operation is invoked, the classes that provides the 
entry-points for Axis (SRMServerV1 and SRMServerV2) resolve the JNDI 
environment to obtain this path, then call SrmDCacheConnector class' 
getInstance method with the patch as an argument.  SrmDCacheConnector 
uses the Singleton anti-pattern, so the first call to getInstance will 
create an instance and subsequent calls will simply return the same 
SrmDCacheConnector instance.

When an SrmDCacheConnector instance is created, it reads the supplied 
path (to srmConfigFile.xml) to discover the name of a class.  Using 
reflection, SrmDCacheConnector calls this class' getSRMInstance method.  
In this case, srmConfigFile.xml specifies that the dCache-specific 
plugin (Storage) provides the getSRMInstance.  This, in turn, relies on 
the singleton-nature of SRM to provide the SRM object.

Once the SrmDCacheConnector is returned, SRMServerV1 and SRMServerV2 
obtain the SRM object directly from SrmDCacheConnector and the plugin 
and configuration are obtained indirectly.

With Jetty and spring, this processes is unnecessarily complicated.  It 
also relies upon various singleton anti-patterns and that objects are 
created in the correct order.

This patch fixes these problems by using named Attributes to transport 
these objects (SRM, Storage, Configuration).  Axis allows access to the 
ServletContext which, in turn, allows access to the Attributes.

With this patch, the SrmDCacheConnector class is unnecessary and the 
singleton nature of SRM is no longer used by SRMServerV1 and 
SRMServerV2 to obtain the three classes.

Note that other parts of the code still depend on SRM class' singleton 
anti-pattern; subsequent patches can remove this.

Target: trunk
Require-book: no
Require-notes: no
Patch: http://rb.dcache.org/r/5078
Acked-by: Dmitry Litvintsev