Add Pool Queue Activity Plots to webadmin
Generate histograms from pool queue load data.

Data collection piggy-backs on the PoolMonitorCollector already running in webadmin.  The collector delegates to an agent for processing round-robin database data and generating plots.

At each pass, the latter extracts six values,

- active movers
- queued movers
- active stores
- queued stores
- active restores
- queued restores

for each pool, and stores them as a time series.  We have adopted rrd4j (see http://code.google.com/p/rrd4j) for this purpose.  The package is used both for data persistence and for graph generation.  Settings for the rrd4j layer are provided via a sub-category of billing,properties, internally set on the RrdSettings object.

Both database (we use the default backend rrd store, named <queue>.rrd) and plot (default is .png) files are written to the httpd.static-content.plots directory (where the billing plots are also stored).

Note that the navigation bar has been adjusted slightly so that pool queue links are adjacent and the billing plots link precedes the strictly administrative links.  The positions are now fixed such that when one is disabled it displays an "Unavailable" page with a message.

Target: trunk
Patch: http://rb.dcache.org/r/5058
Acked-by: Tigran
Require-notes: yes
Require-book: yes

RELEASE NOTES:

Advert the user to the new properties in poolqplots.properties for enabling the collector and setting up the plots.  Once again, with the exception of enabling the collector or not (it is disabled by default), the default properties are sane for most cases so this new functionality will run out of the box.  As for the page itself:

1.  There is a simple filter which works similarly to alarms:  the thumbnails can be sorted up or down alphabetically, or can be filtered using a string or a regex.
2.  The thumbnail is a link which opens a full-size pop-up of the graph.   Not that the table width is determined by the actual number of plots to be displayed by taking the floor of the square root of the number of pools +1, up to a maximum width of 8.  The table size automatically readjusts on filtering as well.  Placeholder icons are used for unoccupied slots in the square when the number of plots is not a perfect square.

Explanation of the various properties is given in poolqplots.properties.

Redeployed trunk and checked for correctness of behavior with webadmin pages.