webdav: Replace HttpManagerBuilder with HttpManagerFactory
HttpManagerBuilder was a modified copy of the HttpManagerBuilder in the Milton library. The code duplication is not desirable.

This patch introduces a custom Spring FactoryBean to inject the custom response handler into the default HttpManagerBuilder two-phase initialization. There is still a minor amount of duplication of the logic in HttpManagerBuilder, but much less than before.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5091/