space-manager: fail requests when disabled
If srmReserveSpace is called with space-manager disabled then dCache 
generates a stack-trace:

15 Jan 2013 17:36:51 (SRM-discordia) [v2:srmGetSpaceMetaData:116868421] handler invocation failed
java.lang.reflect.InvocationTargetException: null
        at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method) ~[na:1.7.0_09]
        at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:57) ~[na:1.7.0_09]
        at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43) ~[na:1.7.0_09]
        at java.lang.reflect.Method.invoke(Method.java:601) ~[na:1.7.0_09]
        at org.dcache.srm.server.SRMServerV2.handleRequest(SRMServerV2.java:214) [srm-2.6.0-SNAPSHOT.jar:2.6.0-SNAPSHOT]
        at org.dcache.srm.server.SRMServerV2.srmGetSpaceMetaData(SRMServerV2.java:305) [srm-2.6.0-SNAPSHOT.jar:2.6.0-SNAPSHOT]
        at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method) ~[na:1.7.0_09]
        at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:57) ~[na:1.7.0_09]
        at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43) ~[na:1.7.0_09]
        at java.lang.reflect.Method.invoke(Method.java:601) ~[na:1.7.0_09]
        at org.apache.axis.providers.java.RPCProvider.invokeMethod(RPCProvider.java:397) [axis-1.4.jar:na]
[... etc ...]
        at org.eclipse.jetty.util.thread.QueuedThreadPool$3.run(QueuedThreadPool.java:533) [jetty-util-7.5.4.v20111024.jar:7.5.4.v20111024]
        at java.lang.Thread.run(Thread.java:722) [na:1.7.0_09]
Caused by: java.lang.NullPointerException: null
        at diskCacheV111.srm.dcache.Storage.srmGetSpaceMetaData(Storage.java:2722) ~[dcache-core-2.6.0-SNAPSHOT.jar:2.6.0-SNAPSHOT]
        at org.dcache.srm.handler.SrmGetSpaceMetaData.srmGetSpaceMetaData(SrmGetSpaceMetaData.java:121) ~[srm-2.6.0-SNAPSHOT.jar:2.6.0-SNAPSHOT]
        at org.dcache.srm.handler.SrmGetSpaceMetaData.getResponse(SrmGetSpaceMetaData.java:75) ~[srm-2.6.0-SNAPSHOT.jar:2.6.0-SNAPSHOT]
        ... 45 common frames omitted

What's happening is that SRM (Storage) is sending a GetSpaceMetaData 
message to Space-Manager.  Space-manager, when "disabled", does not 
write a Space array into the message but throws a SpaceException 
instead.  This is caught and, when disabled, the error-code is set to 
zero (i.e., success).  Note that, without a Spaces array being written 
into the message, getSpaces method will return null.

Unfortunately, the Storage class expects that either an error occurred 
or the Space array contained in the message is non-null.  In this 
particular case, neither is true and the result is a NPE in Storage.

This patch fixes the problem by having Space-Manager return an error if 
called when disabled.  I have a feeling that this may break something 
else, so it would be good for someone with a keen understanding of 
space-manager review this patch.

An alternative approach would be to have the message return an empty 
array rather than null.  This would allow dCache to respond normally, 
as if there are no reservations.

FOR RELEASE NOTES:

Mention that a stack-trace is produced if space-manager operations are 
attempted when space-management is disabled.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.5
Request: 2.4
Request: 2.2
Patch: http://rb.dcache.org/r/5092
Acked-by: Dmitry Litvintsev