chimera: remove SQL file that creates an initial namespace
dCache ships with a SQL file that builds an initial namespace within 
Chimera.

Since chimera schema has been managed via liquibase, this file is 
unnecessary.  Instead, dCache will create the necessary tables and 
rows as part of the database management.  This will happen 
automatically on startup or when the 'dcache database update' command 
is invoked.

In addition, the SQL file contains errors.  The result was that the 
created directories became undeleteable.

Since the file is unnecessary and broken, this patch simply removes 
it.

Note that, for a freshly installed dCache, it is necessary to use the 
'dcache database update' to run the chimera-cli commands before 
starting dCache.

RELEASE NOTES:

mention that create.sql is no longer needed

FOR BOOK:

the install chapter should be updated accordingly.

Target: trunk
Require-book: yes
Require-notes: yes
Patch: http://rb.dcache.org/r/5081/
Acked-by: Tigran Mkrtchyan