Use FileAttributes class in pool and pool related messages
A long time ago we introduced the FileAttributes class. Much code has been
ported to this class, however there is still plenty of code that uses StorageInfo
directly.

This patch updates the pool and pool related messages to use FileAttributes. Care has
been taken to maintain backwards compatibility with older pools.

StorageInfo is still used, and will likely also be used in the future: The
pool serializes this class and thus it will not be easy to get rid of. Our code
should however as far as possible use FileAttributes.

Now is a perfect time for the patch. The code in the messages that maintains
backwards compatibility can be removed between the release of 2.6 and 2.7.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5107/